# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 3 Batti Betting Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with MPIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpBody"
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-mpin:
    post:
      operationId: resetMpin
      tags: [auth]
      summary: Reset MPIN via OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetMpinBody"
      responses:
        "200":
          description: MPIN reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/refresh-token:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenBody"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  # ADMIN AUTH
  /auth/admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Admin logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  # USER
...
  /game/results:
    get:
      operationId: getGameResults
      tags: [game]
      summary: Get last 50 results
      parameters:
        - name: page
          in: query
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: Results list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameResultList"
  /game/place-bet:
    post:
      operationId: placeBet
      tags: [game]
      summary: Place a bet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceBetBody"
      responses:
        "201":
          description: Bet placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bet"
  /user/bets:
    get:
      operationId: getUserBets
      tags: [game]
      summary: Get user bet history
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
          required: false
        - name: gameId
          in: query
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: Bet history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BetList"

  # PAYMENT
  /payment/deposit-request:
    post:
      operationId: createDepositRequest
      tags: [payment]
      summary: Submit deposit request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositRequestBody"
      responses:
        "201":
          description: Deposit request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /payment/withdrawal-request:
    post:
      operationId: createWithdrawalRequest
      tags: [payment]
      summary: Submit withdrawal request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalRequestBody"
      responses:
        "201":
          description: Withdrawal request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /payment/upi-details:
    get:
      operationId: getUpiDetails
      tags: [payment]
      summary: Get UPI/QR assigned to user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: UPI details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpiDetails"

  # REFERRAL
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
      required: [message, success]

    # AUTH
    RegisterBody:
      type: object
      properties:
        mobile:
          type: string
        name:
          type: string
        mpin:
          type: string
        referralCode:
          type: string
      required: [mobile, name, mpin]
    LoginBody:
      type: object
      properties:
        mobile:
          type: string
        mpin:
          type: string
      required: [mobile, mpin]
    AdminLoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    VerifyOtpBody:
      type: object
      properties:
        mobile:
          type: string
        otp:
          type: string
      required: [mobile, otp]
    ResetMpinBody:
      type: object
      properties:
        mobile:
          type: string
        otp:
          type: string
        newMpin:
          type: string
      required: [mobile, otp, newMpin]
    RefreshTokenBody:
      type: object
      properties:
        refreshToken:
          type: string
      required: [refreshToken]
...
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [accessToken, refreshToken, user]

    # USER
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        mobile:
          type: string
        email:
          type: string
        referralCode:
          type: string
        kycStatus:
          type: string
          enum: [pending, verified]
        isBlocked:
          type: boolean
        walletBalance:
          type: number
        createdAt:
          type: string
      required: [id, name, mobile, referralCode, kycStatus, isBlocked, walletBalance, createdAt]
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, bet, win, referral, bonus]
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected, processing, completed]
        referenceId:
          type: string
        screenshotUrl:
          type: string
        reason:
          type: string
        createdAt:
          type: string
        userName:
          type: string
        userMobile:
          type: string
      required: [id, userId, type, amount, status, createdAt]
...
    GameResult:
      type: object
      properties:
        id:
          type: integer
        scheduledAt:
          type: string
        status:
          type: string
          enum: [upcoming, live, closed]
        winningColor:
          type: string
          enum: [red, yellow, green]
        winningNumber:
          type: integer
        totalCollection:
          type: number
        totalPayout:
          type: number
        createdAt:
          type: string
      required: [id, scheduledAt, status, createdAt]
...
    GameResultList:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/GameResult"
        total:
          type: integer
      required: [results, total]
...
    Bet:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        gameId:
          type: integer
        betType:
          type: string
          enum: [color, number]
        selection:
          type: string
        amount:
          type: number
        potentialWin:
          type: number
        status:
          type: string
          enum: [pending, won, lost]
        createdAt:
          type: string
      required: [id, userId, gameId, betType, selection, amount, potentialWin, status, createdAt]
    BetList:
      type: object
      properties:
        bets:
          type: array
          items:
            $ref: "#/components/schemas/Bet"
        total:
          type: integer
        page:
          type: integer
      required: [bets, total, page]
    PlaceBetBody:
      type: object
      properties:
        gameId:
          type: integer
        betType:
          type: string
          enum: [color, number]
        selection:
          type: string
        amount:
          type: number
      required: [gameId, betType, selection, amount]

    # PAYMENT
    DepositRequestBody:
      type: object
      properties:
        amount:
          type: number
        upiId:
          type: string
        referenceId:
          type: string
        screenshotUrl:
          type: string
      required: [amount, referenceId]
    WithdrawalRequestBody:
      type: object
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [bank, upi]
        upiId:
          type: string
      required: [amount, method]
    UpiDetails:
      type: object
      properties:
        upiId:
          type: string
        qrImageUrl:
          type: string
        displayName:
          type: string
      required: [upiId]
...